Feat: Added SubMenuUserData.jsx and .css in ticket_buy folder for this section of the buying ticket proccess. (#37)
Also refactores CampoTexto so that it can receive an action to do when changing the input value, so parents can recover it and use it.

Added basic funcionality of the SubMenu, all input data is saved in a so called state in json format so that it is easy to implement later the sending of the data to the api server.