Feature: Added field checking for user data in ticket buy page.
Also added error codes and messages to report feedback about potential errors.

With that purpose added a new Alert component meant to display this type of info to the user.

Had to partially refactor the ticket buy app, simplifying the responsiblity of the main.jsx and delegating to the app.jsx.

TODO: send data to api

TODO: receive data from api and render the correct event