Refactor: api methods in ticket buy
Modified api functions in ticket buy so that they can ether fetch from a given server or fetch locally some files.

With that purpose added a .env to have an environment variable set to easily configure local or production environment

Modified the sign of some methods

Made methods in app async to potentially wait for server response

Modified app to be consistent with new method signs.